moved UI views to new application "ui" to avoid a "soft" circular reference.
