+++ smswebapp/settings/base.py
@@ -34,6 +34,7 @@ INSTALLED_APPS = [
 
     'smsjwplatform',
     'legacysms',
+    'ui',
 ]
 
 #: Installed middleware
