+++ ui/__init__.py
@@ -1,0 +1,4 @@
+"""
+The :py:mod:`ui` application provides the University Media Service UI.
+
+"""
