+++ ui/tests/test_views.py
@@ -1,0 +1,13 @@
+"""
+Tests for views.
+
+"""
+
+# No views at present
+from unittest import TestCase
+
+
+class ViewsTestCase(TestCase):
+
+    def test_it(self):
+        self.fail()
